RAG/AI-product/requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
google-genai
numpy
pandas
